+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 * `[New]` Optional functionality to change user password on Active Directory via LDAP.
 * `[Fix]` Update password complexity checks to accept non-7-bit ASCII.
 * `[Fix]` Fix 2FA secrets 160-bit key issue.
+* `[Fix]` Fix issue with LDAP password change for users that have never logged in before.